App: Extracted common `didMetaQueryBase`
and placed it into the root folder of `app/did` so that it may be
accessed by the `page` and `list` subpages.
